Remove login page
