+++ frontend/src/components/layouts/MainLayout.tsx
@@ -1,11 +1,9 @@
 import { Stack } from "@mui/material";
 import { Outlet } from "react-router-dom";
-import MainHeader from "../headers/MainHeader";
 
 function MainLayout() {
 	return (
 		<Stack sx={{ flexGrow: 1, height: "100vh" }}>
-			<MainHeader />
 			<Outlet />
 		</Stack>
 	);
